// generated/interface/models/InterfaceInfo.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type InterfaceInfo = {
    createTime?: string;
    description?: string;
    id?: number;
    isDelete?: number;
    method?: string;
    name?: string;
    requestParam?: string;
    responseBody?: string;
    status?: number;
    updateTime?: string;
    url?: string;
    userId?: number;
};